+++ backend/coffeeing/src/main/java/com/ssafy/coffeeing/modules/product/service/CoffeeReviewService.java
@@ -86,9 +86,9 @@ public class CoffeeReviewService {
             throw new BusinessException(AuthErrorInfo.UNAUTHORIZED);
         }
 
-        review.getCoffee().editReview(reviewRequest.score() - (int) (double) review.getCoffee().getTotalScore());
+        review.getCoffee().editReview(reviewRequest.score() - review.getCoffee().getTotalScore());
 
-        review.update(reviewRequest.content(), (double) reviewRequest.score());
+        review.update(reviewRequest.content(), reviewRequest.score());
     }
 
     public void deleteReview(Long id) {
